+++ opentelemetry-sdk/src/opentelemetry/sdk/_logs/_internal/export/in_memory_log_exporter.py
@@ -46,7 +46,7 @@ class InMemoryLogRecordExporter(LogRecordExporter):
     def shutdown(self) -> None:
         self._stopped = True
 
-    def force_flush(self, timeout_millis: float = 10_000) -> bool:
+    def force_flush(self, timeout_millis: int = 10_000) -> bool:
         return True
